+++ src/generators/generators/piecewise-merge-join-node.generator.ts
@@ -194,7 +194,8 @@ export class PiecewiseMergeJoinNodeGenerator extends BaseNodeGenerator {
       startPositions[outerIndex] ?? childInfo.x + childInfo.width / 2,
       childY,
       endPositions[outerIndex] ?? (endLeft + endRight) / 2,
-      parentBottomY
+      parentBottomY,
+      childInfo.groupId
     );
   }